backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pillow
google-generativeai
python-dotenv
requests
